requirements.txt
streamlit>=1.28.0
selenium>=4.15.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.2
pandas>=2.1.2
tenacity>=8.2.3
requests>=2.31.0
Pillow>=10.1.0
validators>=0.22.0
ratelimit>=2.2.1
fake-useragent>=1.3.0
newspaper3k>=0.2.8
nltk>=3.8.1
aiohttp>=3.8.6
PyYAML>=6.0.1
lxml>=4.9.3
python-dateutil>=2.8.2
urllib3>=2.0.7
certifi>=2023.7.22
charset-normalizer>=3.3.1
idna>=3.4
six>=1.16.0
pytz>=2023.3.post1
numpy>=1.26.1
python-dotenv>=1.0.0
tqdm>=4.66.1
feedparser>=6.0.10
cssselect>=1.2.0
jieba3k>=0.35.1
feedfinder2>=0.0.4
sgmllib3k>=1.0.0
tinysegmenter>=0.3
requests-file>=1.5.1
pytest>=7.4.3
black>=23.10.1
flake8>=6.1.0
mypy>=1.6.1
isort>=5.12.0
html5lib>=1.1
soupsieve>=2.5
multidict>=6.0.4
async-timeout>=4.0.3
frozenlist>=1.4.0
yarl>=1.9.2
attrs>=23.1.0
packaging>=23.2
pluggy>=1.3.0
iniconfig>=2.0.0
exceptiongroup>=1.1.3
tomli>=2.0.1
typing-extensions>=4.8.0
mypy-extensions>=1.0.0
pathspec>=0.11.2
mccabe>=0.7.0
pycodestyle>=2.11.1
pyflakes>=3.1.0
webencodings>=0.5.1
cryptography>=41.0.4
pyOpenSSL>=23.2.0
cffi>=1.16.0
pycparser>=2.21
streamlit-aggrid>=0.3.4.post3
streamlit-option-menu>=0.3.6
streamlit-elements>=0.1.0
cachetools>=5.3.2
joblib>=1.3.2
spacy>=3.0.0
en_core_web_sm
undetected_chromedriver
textblob
twocaptcha